Move files to parent folder due to compilation demands
