+++ Artist.cpp
