+++ ArtistPlays.h
@@ -5,8 +5,8 @@
 #ifndef WET1_ARTISTPLAYS_H
 #define WET1_ARTISTPLAYS_H
 
-#include "../TreeNode/Tree.h"
-#include "../SongPlays/SongPlays.h"
+#include "Tree.h"
+#include "SongPlays.h"
 
 class MostPlayedListNode;
 class SongPlays;
